scrapping script/requirements.txt
python-dotenv==1.0.1
pandas==2.2.1
selenium==4.18.1
webdriver-manager==4.0.1
